const 형식 제거 후 디폴트 대입연산자로 호출

+++ IOCP/IOCP_SERVER/Server/RoomManager/RoomManager.cpp
@@ -25,7 +25,7 @@ void RoomManager::EnterRoom(int roomNum, ClientInfo* clientInfo)
 
 int RoomManager::GetRoomVecSize()
 {
-	return _roomVec.size();
+	return static_cast<int>(_roomVec.size());
 }
 
 Room RoomManager::GetRoomInfoByRoomNum(int roomNum)
